fix: first phase of bug fixes
