+++ app/src/main/java/com/rishav/gidget/UI/MainActivity.kt
@@ -37,7 +37,7 @@ class MainActivity : AppCompatActivity() {
 
     private fun loginButtonOnTap() {
         val progressBar: ProgressBar = findViewById(R.id.mainPageProgressBar)
-        val loginButton: Button = findViewById<Button>(R.id.buLogin)
+        val loginButton: Button = findViewById(R.id.buLogin)
         loginButton.setOnClickListener {
             progressBar.visibility = View.VISIBLE
             loginButton.visibility = View.GONE
@@ -56,7 +56,11 @@ class MainActivity : AppCompatActivity() {
     }
 
     @Suppress("RECEIVER_NULLABILITY_MISMATCH_BASED_ON_JAVA_ANNOTATIONS")
-    private fun newLogin(provider: OAuthProvider.Builder, progressBar: ProgressBar, loginButton: Button) {
+    private fun newLogin(
+        provider: OAuthProvider.Builder,
+        progressBar: ProgressBar,
+        loginButton: Button
+    ) {
         mAuth!!.startActivityForSignInWithProvider(this, provider.build())
             .addOnSuccessListener {
                 Realm.init(applicationContext)
@@ -91,7 +95,7 @@ class MainActivity : AppCompatActivity() {
             .addOnFailureListener {
                 progressBar.visibility = View.GONE
                 loginButton.visibility = View.VISIBLE
-                Toast.makeText(this, it.toString(), Toast.LENGTH_SHORT).show()
+                Toast.makeText(this, "Login error", Toast.LENGTH_SHORT).show()
                 println(it)
             }
     }
@@ -99,14 +103,14 @@ class MainActivity : AppCompatActivity() {
     private fun pendingLogin(progressBar: ProgressBar, loginButton: Button) {
         val pendingResultTask: Task<AuthResult> = mAuth!!.pendingAuthResult!!
         pendingResultTask.addOnSuccessListener {
-            Toast.makeText(this, "Logged in", Toast.LENGTH_LONG).show()
+            Toast.makeText(this, "Pending Logged in", Toast.LENGTH_LONG).show()
             startActivity(Intent(this, FeedActivity::class.java))
             finish()
         }
             .addOnFailureListener {
                 progressBar.visibility = View.GONE
                 loginButton.visibility = View.VISIBLE
-                Toast.makeText(this, it.toString(), Toast.LENGTH_SHORT).show()
+                Toast.makeText(this, "Pending login error", Toast.LENGTH_SHORT).show()
                 println(it)
             }
     }
