--- app/src/main/res/drawable/appwidget_outer_bg.xml
@@ -1,5 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<shape xmlns:android="http://schemas.android.com/apk/res/android">
-    <solid android:color="@color/black"/>
-    <corners android:radius="10dp"/>
-</shape>
