+++ app/src/main/res/drawable/ic_git_branch.xml
@@ -3,8 +3,8 @@
     android:height="16dp"
     android:viewportWidth="16"
     android:viewportHeight="16">
-  <path
-      android:pathData="M11.8,2.5c-0.4,0 -0.8,0.3 -0.8,0.8S11.3,4 11.8,4s0.8,-0.3 0.8,-0.8S12.2,2.5 11.8,2.5zM9.5,3.3c0,-1.2 1,-2.2 2.3,-2.2C13,1 14,2 14,3.3c0,1 -0.6,1.8 -1.5,2.1V6c0,1.4 -1.1,2.5 -2.5,2.5H6c-0.6,0 -1,0.4 -1,1v1.1c1.2,0.4 1.8,1.7 1.4,2.9c-0.4,1.2 -1.7,1.8 -2.9,1.4c-1.2,-0.4 -1.8,-1.7 -1.4,-2.9c0.2,-0.6 0.7,-1.1 1.4,-1.4V5.4C2.3,5 1.7,3.7 2.1,2.5S3.8,0.7 5,1.1S6.8,2.8 6.4,4C6.1,4.6 5.6,5.1 5,5.4v1.8C5.3,7.1 5.7,7 6,7h4c0.6,0 1,-0.4 1,-1V5.4C10.1,5.1 9.5,4.2 9.5,3.3zM4.3,12c-0.4,0 -0.8,0.3 -0.8,0.8s0.3,0.8 0.8,0.8S5,13.2 5,12.8S4.7,12 4.3,12zM3.5,3.3c0,-0.4 0.3,-0.8 0.8,-0.8S5,2.8 5,3.3S4.7,4 4.3,4S3.5,3.7 3.5,3.3z"
-      android:fillColor="#FFFFFF"
-      android:fillType="evenOdd"/>
+    <path
+        android:fillColor="@color/blue"
+        android:fillType="evenOdd"
+        android:pathData="M11.8,2.5c-0.4,0 -0.8,0.3 -0.8,0.8S11.3,4 11.8,4s0.8,-0.3 0.8,-0.8S12.2,2.5 11.8,2.5zM9.5,3.3c0,-1.2 1,-2.2 2.3,-2.2C13,1 14,2 14,3.3c0,1 -0.6,1.8 -1.5,2.1V6c0,1.4 -1.1,2.5 -2.5,2.5H6c-0.6,0 -1,0.4 -1,1v1.1c1.2,0.4 1.8,1.7 1.4,2.9c-0.4,1.2 -1.7,1.8 -2.9,1.4c-1.2,-0.4 -1.8,-1.7 -1.4,-2.9c0.2,-0.6 0.7,-1.1 1.4,-1.4V5.4C2.3,5 1.7,3.7 2.1,2.5S3.8,0.7 5,1.1S6.8,2.8 6.4,4C6.1,4.6 5.6,5.1 5,5.4v1.8C5.3,7.1 5.7,7 6,7h4c0.6,0 1,-0.4 1,-1V5.4C10.1,5.1 9.5,4.2 9.5,3.3zM4.3,12c-0.4,0 -0.8,0.3 -0.8,0.8s0.3,0.8 0.8,0.8S5,13.2 5,12.8S4.7,12 4.3,12zM3.5,3.3c0,-0.4 0.3,-0.8 0.8,-0.8S5,2.8 5,3.3S4.7,4 4.3,4S3.5,3.7 3.5,3.3z" />
 </vector>
