+++ app/src/main/res/drawable/ic_github_fork.xml
@@ -5,6 +5,6 @@
     android:viewportHeight="39">
   <path
       android:pathData="M8.244,6.2616C8.244,7.4038 7.3089,8.351 6.1814,8.351C5.0538,8.351 4.1187,7.4038 4.1187,6.2616C4.1187,5.1194 5.0538,4.1722 6.1814,4.1722C7.3089,4.1722 8.244,5.1194 8.244,6.2616ZM8.244,12.1676C11.4617,11.0254 13.1668,7.4317 12.0118,4.1722C10.8567,0.9127 7.3365,-0.7867 4.1187,0.3555C0.901,1.4977 -0.7766,5.0915 0.351,8.351C0.9835,10.134 2.3586,11.5548 4.1187,12.1676V14.6192C4.1187,18.0737 6.8964,20.8874 10.3067,20.8874H14.4319V26.8214C11.2142,27.9636 9.5091,31.5574 10.6642,34.8168C11.7918,38.0763 15.3395,39.8035 18.5572,38.6335C21.7749,37.4913 23.48,33.8975 22.325,30.638C21.7199,28.855 20.3173,27.4343 18.5572,26.8214V20.8874H22.6825C26.0927,20.8874 28.8704,18.0737 28.8704,14.6192V12.1676C32.0881,11.0254 33.7933,7.4317 32.6382,4.1722C31.5106,0.9127 27.9629,-0.7867 24.7451,0.3555C21.5274,1.4977 19.8498,5.0915 20.9774,8.351C21.6099,10.134 22.985,11.5548 24.7451,12.1676V14.6192C24.7451,15.7614 23.8101,16.7086 22.6825,16.7086H10.3067C9.1791,16.7086 8.244,15.7614 8.244,14.6192V12.1676ZM18.5572,32.7274C18.5572,33.8696 17.6221,34.8168 16.4946,34.8168C15.367,34.8168 14.4319,33.8696 14.4319,32.7274C14.4319,31.5852 15.367,30.638 16.4946,30.638C17.6221,30.638 18.5572,31.5852 18.5572,32.7274ZM26.8078,8.351C27.9354,8.351 28.8704,7.4038 28.8704,6.2616C28.8704,5.1194 27.9354,4.1722 26.8078,4.1722C25.6802,4.1722 24.7451,5.1194 24.7451,6.2616C24.7451,7.4038 25.6802,8.351 26.8078,8.351Z"
-      android:fillColor="#ffffff"
+      android:fillColor="@color/blue"
       android:fillType="evenOdd"/>
 </vector>
