+++ app/src/main/res/drawable/ic_github_issue.xml
@@ -1,10 +1,12 @@
 <vector xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:tools="http://schemas.android.com/tools"
     android:width="41dp"
     android:height="41dp"
     android:viewportWidth="41"
     android:viewportHeight="41">
-  <path
-      android:pathData="M20.5,3.8438C11.3006,3.8438 3.8438,11.3006 3.8438,20.5C3.8438,29.6994 11.3006,37.1562 20.5,37.1562C29.6994,37.1562 37.1562,29.6994 37.1562,20.5C37.1562,11.3006 29.6994,3.8438 20.5,3.8438ZM0,20.5C0,9.1737 9.1737,0 20.5,0C31.8263,0 41,9.1737 41,20.5C41,31.8263 31.8263,41 20.5,41C9.1737,41 0,31.8263 0,20.5ZM23.0625,28.1875C23.0625,29.5969 21.9094,30.75 20.5,30.75C19.0906,30.75 17.9375,29.5969 17.9375,28.1875C17.9375,26.7781 19.0906,25.625 20.5,25.625C21.9094,25.625 23.0625,26.7781 23.0625,28.1875ZM22.4219,12.1719C22.4219,11.1213 21.5506,10.25 20.5,10.25C19.4494,10.25 18.5781,11.1213 18.5781,12.1719V21.1406C18.5781,22.1912 19.4494,23.0625 20.5,23.0625C21.5506,23.0625 22.4219,22.1912 22.4219,21.1406V12.1719Z"
-      android:fillColor="#ffffff"
-      android:fillType="evenOdd"/>
+    <path
+        android:fillColor="@color/blue"
+        android:fillType="evenOdd"
+        android:pathData="M20.5,3.8438C11.3006,3.8438 3.8438,11.3006 3.8438,20.5C3.8438,29.6994 11.3006,37.1562 20.5,37.1562C29.6994,37.1562 37.1562,29.6994 37.1562,20.5C37.1562,11.3006 29.6994,3.8438 20.5,3.8438ZM0,20.5C0,9.1737 9.1737,0 20.5,0C31.8263,0 41,9.1737 41,20.5C41,31.8263 31.8263,41 20.5,41C9.1737,41 0,31.8263 0,20.5ZM23.0625,28.1875C23.0625,29.5969 21.9094,30.75 20.5,30.75C19.0906,30.75 17.9375,29.5969 17.9375,28.1875C17.9375,26.7781 19.0906,25.625 20.5,25.625C21.9094,25.625 23.0625,26.7781 23.0625,28.1875ZM22.4219,12.1719C22.4219,11.1213 21.5506,10.25 20.5,10.25C19.4494,10.25 18.5781,11.1213 18.5781,12.1719V21.1406C18.5781,22.1912 19.4494,23.0625 20.5,23.0625C21.5506,23.0625 22.4219,22.1912 22.4219,21.1406V12.1719Z"
+        tools:ignore="VectorRaster" />
 </vector>
