+++ app/src/main/res/drawable/ic_github_pull_request.xml
@@ -1,10 +1,12 @@
 <vector xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:tools="http://schemas.android.com/tools"
     android:width="40dp"
     android:height="42dp"
     android:viewportWidth="40"
     android:viewportHeight="42">
-  <path
-      android:pathData="M16.8304,7.2155L23.9398,0.2186C24.2361,-0.0729 24.71,-0.0729 24.9766,0.2186C25.1247,0.3644 25.184,0.5393 25.184,0.7434V14.7079C25.184,15.116 24.8581,15.4367 24.4434,15.4367C24.2361,15.4367 24.0583,15.3493 23.9102,15.2326L16.8007,8.2358C16.5341,7.9735 16.5341,7.507 16.8304,7.2155ZM6.6698,5.5537C5.4552,5.5537 4.4481,6.5158 4.4481,7.7402C4.4481,8.9647 5.4552,9.9267 6.6698,9.9267C7.8843,9.9267 8.8915,8.9355 8.8915,7.7402C8.8915,6.5449 7.9139,5.5537 6.6698,5.5537ZM0.0046,7.7402C0.0046,4.096 2.9965,1.1807 6.6698,1.1807C10.343,1.1807 13.3349,4.1252 13.3349,7.7402C13.3349,10.5098 11.5575,12.9878 8.8915,13.9207V29.2554C12.3573,30.4507 14.1939,34.2115 12.9498,37.6224C11.7056,41.0334 7.9139,42.8409 4.4481,41.6164C0.9822,40.4212 -0.8544,36.6604 0.3897,33.2494C1.0711,31.3836 2.5522,29.8968 4.4481,29.2554V13.9207C1.782,12.9878 0.0046,10.5098 0.0046,7.7402ZM28.1463,5.5537H25.184V9.9267H28.1463C29.7755,9.9267 31.1085,11.2386 31.1085,12.8421V29.2554C27.6427,30.4507 25.8061,34.2115 27.0502,37.6224C28.2648,41.0334 32.0861,42.8409 35.552,41.6164C39.0178,40.4212 40.8544,36.6604 39.6103,33.2494C38.9289,31.3836 37.4478,29.8968 35.552,29.2554V12.8421C35.552,8.8189 32.2342,5.5537 28.1463,5.5537ZM31.1085,35.4359C31.1085,34.2406 32.1157,33.2494 33.3302,33.2494C34.5448,33.2494 35.552,34.2115 35.552,35.4359C35.552,36.6604 34.5448,37.6224 33.3302,37.6224C32.1157,37.6224 31.1085,36.6312 31.1085,35.4359ZM6.6698,33.2494C5.4552,33.2494 4.4481,34.2406 4.4481,35.4359C4.4481,36.6312 5.4552,37.6224 6.6698,37.6224C7.8843,37.6224 8.8915,36.6312 8.8915,35.4359C8.8915,34.2406 7.9139,33.2494 6.6698,33.2494Z"
-      android:fillColor="#ffffff"
-      android:fillType="evenOdd"/>
+    <path
+        android:fillColor="@color/blue"
+        android:fillType="evenOdd"
+        android:pathData="M16.8304,7.2155L23.9398,0.2186C24.2361,-0.0729 24.71,-0.0729 24.9766,0.2186C25.1247,0.3644 25.184,0.5393 25.184,0.7434V14.7079C25.184,15.116 24.8581,15.4367 24.4434,15.4367C24.2361,15.4367 24.0583,15.3493 23.9102,15.2326L16.8007,8.2358C16.5341,7.9735 16.5341,7.507 16.8304,7.2155ZM6.6698,5.5537C5.4552,5.5537 4.4481,6.5158 4.4481,7.7402C4.4481,8.9647 5.4552,9.9267 6.6698,9.9267C7.8843,9.9267 8.8915,8.9355 8.8915,7.7402C8.8915,6.5449 7.9139,5.5537 6.6698,5.5537ZM0.0046,7.7402C0.0046,4.096 2.9965,1.1807 6.6698,1.1807C10.343,1.1807 13.3349,4.1252 13.3349,7.7402C13.3349,10.5098 11.5575,12.9878 8.8915,13.9207V29.2554C12.3573,30.4507 14.1939,34.2115 12.9498,37.6224C11.7056,41.0334 7.9139,42.8409 4.4481,41.6164C0.9822,40.4212 -0.8544,36.6604 0.3897,33.2494C1.0711,31.3836 2.5522,29.8968 4.4481,29.2554V13.9207C1.782,12.9878 0.0046,10.5098 0.0046,7.7402ZM28.1463,5.5537H25.184V9.9267H28.1463C29.7755,9.9267 31.1085,11.2386 31.1085,12.8421V29.2554C27.6427,30.4507 25.8061,34.2115 27.0502,37.6224C28.2648,41.0334 32.0861,42.8409 35.552,41.6164C39.0178,40.4212 40.8544,36.6604 39.6103,33.2494C38.9289,31.3836 37.4478,29.8968 35.552,29.2554V12.8421C35.552,8.8189 32.2342,5.5537 28.1463,5.5537ZM31.1085,35.4359C31.1085,34.2406 32.1157,33.2494 33.3302,33.2494C34.5448,33.2494 35.552,34.2115 35.552,35.4359C35.552,36.6604 34.5448,37.6224 33.3302,37.6224C32.1157,37.6224 31.1085,36.6312 31.1085,35.4359ZM6.6698,33.2494C5.4552,33.2494 4.4481,34.2406 4.4481,35.4359C4.4481,36.6312 5.4552,37.6224 6.6698,37.6224C7.8843,37.6224 8.8915,36.6312 8.8915,35.4359C8.8915,34.2406 7.9139,33.2494 6.6698,33.2494Z"
+        tools:ignore="VectorRaster" />
 </vector>
