+++ app/src/main/res/drawable/login_button_image.xml
@@ -1,9 +1,10 @@
 <?xml version="1.0" encoding="utf-8"?>
-<layer-list xmlns:tools="http://schemas.android.com/tools"
-    xmlns:android="http://schemas.android.com/apk/res/android">
+<layer-list xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:tools="http://schemas.android.com/tools">
     <item
         android:width="30dp"
         android:height="30dp"
         android:drawable="@drawable/github_logo"
-        tools:targetApi="m" />
+        tools:targetApi="m"
+        tools:ignore="UnusedAttribute" />
 </layer-list>
