+++ app/src/main/res/layout/activity_feed.xml
@@ -28,14 +28,16 @@
                     android:layout_width="wrap_content"
                     android:layout_height="wrap_content"
                     android:layout_alignParentStart="true"
-                    app:cardCornerRadius="50dp">
+                    android:shape="ring"
+                    app:cardCornerRadius="@dimen/_14sdp">
 
                     <ImageView
                         android:id="@+id/feedPageProfilePhoto"
-                        android:layout_width="@dimen/_25sdp"
-                        android:layout_height="@dimen/_25sdp"
+                        android:layout_width="@dimen/_28sdp"
+                        android:layout_height="@dimen/_28sdp"
                         android:background="?android:attr/selectableItemBackground"
                         android:clickable="true"
+                        android:contentDescription="@string/picture"
                         android:focusable="true"
                         android:src="@drawable/github_logo" />
                 </androidx.cardview.widget.CardView>
@@ -48,7 +50,8 @@
                     android:fontFamily="@font/nexa_bold"
                     android:text="@string/app_name"
                     android:textColor="@color/white"
-                    android:textSize="@dimen/_20sdp" />
+                    android:textSize="@dimen/_20sdp"
+                    tools:ignore="RelativeOverlap" />
 
                 <androidx.cardview.widget.CardView
                     android:id="@+id/feedPageSearchButton"
@@ -59,8 +62,9 @@
                     android:clickable="true"
                     android:focusable="true"
                     android:foreground="?android:attr/selectableItemBackground"
+                    android:shape="ring"
                     app:cardBackgroundColor="@color/feedPageRecyclerItemColor"
-                    app:cardCornerRadius="30dp">
+                    app:cardCornerRadius="@dimen/_15sdp">
 
                     <ImageView
                         android:layout_width="@dimen/_20sdp"
@@ -84,7 +88,7 @@
         android:layout_marginHorizontal="30dp"
         android:layout_marginVertical="@dimen/_20sdp"
         android:fontFamily="@font/montserrat"
-        android:text="Activity"
+        android:text="@string/activity"
         android:textColor="@color/white"
         android:textSize="24sp" />
 
@@ -94,6 +98,7 @@
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:layout_below="@+id/feedPageActvityTV"
+        android:scrollbars="vertical"
         tools:listitem="@layout/feed_page_recycler_item" />
 
     <RelativeLayout
@@ -102,6 +107,10 @@
         android:layout_height="match_parent"
         android:layout_alignParentStart="true"
         android:layout_alignParentTop="true"
+        android:layout_alignParentBottom="true"
+        android:layout_marginStart="2dp"
+        android:layout_marginTop="-19dp"
+        android:layout_marginBottom="0dp"
         android:gravity="center">
 
         <ProgressBar
@@ -111,6 +120,7 @@
             android:indeterminateTint="@color/white"
             android:minWidth="@dimen/_50sdp"
             android:minHeight="@dimen/_50sdp"
+            android:visibility="gone"
             tools:targetApi="lollipop" />
     </RelativeLayout>
 </RelativeLayout>
