+++ app/src/main/res/layout/activity_profile.xml
@@ -32,6 +32,7 @@
                     android:layout_marginEnd="10dp"
                     android:background="?android:attr/selectableItemBackground"
                     android:clickable="true"
+                    android:contentDescription="@string/picture"
                     android:focusable="true"
                     android:src="@drawable/ic_baseline_arrow_back_24" />
 
@@ -69,13 +70,14 @@
                 android:layout_height="wrap_content"
                 android:layout_alignParentStart="true"
                 app:cardBackgroundColor="@android:color/transparent"
-                app:cardCornerRadius="100dp"
+                app:cardCornerRadius="@dimen/_40sdp"
                 app:cardElevation="0dp">
 
                 <ImageView
                     android:id="@+id/profilePagePhoto"
-                    android:layout_width="100dp"
-                    android:layout_height="100dp"
+                    android:layout_width="@dimen/_80sdp"
+                    android:layout_height="@dimen/_80sdp"
+                    android:contentDescription="@string/picture"
                     android:src="@drawable/ic_baseline_person_24" />
             </androidx.cardview.widget.CardView>
 
@@ -92,6 +94,7 @@
                     android:layout_width="wrap_content"
                     android:layout_height="wrap_content"
                     android:layout_gravity="end"
+                    android:layout_marginBottom="@dimen/_5sdp"
                     android:fontFamily="@font/montserrat"
                     android:text="@string/name"
                     android:textAlignment="textEnd"
@@ -125,13 +128,14 @@
                 android:layout_width="wrap_content"
                 android:layout_height="wrap_content"
                 android:layout_alignParentStart="true"
+                android:layout_marginEnd="@dimen/_6sdp"
                 android:orientation="horizontal">
 
                 <TextView
                     android:id="@+id/profilePageFollowerCount"
                     android:layout_width="wrap_content"
                     android:layout_height="wrap_content"
-                    android:text="0"
+                    android:text="@string/numbersCount"
                     android:textColor="@color/blue"
                     android:textSize="20sp" />
 
@@ -140,7 +144,7 @@
                     android:layout_height="wrap_content"
                     android:layout_marginStart="10dp"
                     android:fontFamily="@font/montserrat"
-                    android:text="Followers"
+                    android:text="@string/followers"
                     android:textColor="@color/white"
                     android:textSize="20sp" />
             </LinearLayout>
@@ -149,13 +153,15 @@
                 android:layout_width="wrap_content"
                 android:layout_height="wrap_content"
                 android:layout_alignParentEnd="true"
-                android:orientation="horizontal">
+                android:layout_marginStart="@dimen/_6sdp"
+                android:orientation="horizontal"
+                tools:ignore="RelativeOverlap">
 
                 <TextView
                     android:id="@+id/profilePageFollowingCount"
                     android:layout_width="wrap_content"
                     android:layout_height="wrap_content"
-                    android:text="0"
+                    android:text="@string/numbersCount"
                     android:textColor="@color/blue"
                     android:textSize="20sp" />
 
@@ -164,7 +170,7 @@
                     android:layout_height="wrap_content"
                     android:layout_marginStart="10dp"
                     android:fontFamily="@font/montserrat"
-                    android:text="Following"
+                    android:text="@string/following"
                     android:textColor="@color/white"
                     android:textSize="20sp" />
             </LinearLayout>
@@ -183,7 +189,7 @@
                 android:layout_width="wrap_content"
                 android:layout_height="wrap_content"
                 android:fontFamily="@font/montserrat"
-                android:text="Bio"
+                android:text="@string/bio"
                 android:textColor="@color/white"
                 android:textSize="20sp"
                 android:textStyle="bold" />
@@ -206,11 +212,13 @@
             android:layout_below="@+id/profilePageSection3"
             android:layout_marginHorizontal="30dp"
             android:layout_marginVertical="15dp"
-            android:orientation="horizontal">
+            android:orientation="horizontal"
+            tools:ignore="UseCompoundDrawables">
 
             <ImageView
                 android:layout_width="30dp"
                 android:layout_height="30dp"
+                android:contentDescription="@string/picture"
                 android:src="@drawable/ic_baseline_location_on_24" />
 
             <TextView
@@ -220,7 +228,7 @@
                 android:layout_gravity="center"
                 android:layout_marginStart="10dp"
                 android:fontFamily="@font/montserrat"
-                android:text="City"
+                android:text="@string/city"
                 android:textColor="@color/white"
                 android:textSize="20sp" />
         </LinearLayout>
@@ -232,20 +240,21 @@
             android:layout_alignParentBottom="true"
             android:layout_marginHorizontal="30dp"
             android:layout_marginBottom="30dp"
-            android:foreground="?android:attr/selectableItemBackground"
             android:clickable="true"
             android:focusable="true"
+            android:foreground="?android:attr/selectableItemBackground"
             app:cardBackgroundColor="@color/feedPageRecyclerItemColor"
             app:cardCornerRadius="10dp"
             app:cardElevation="0dp">
 
             <TextView
+                android:id="@+id/profilePageLogoutButtonText"
                 android:layout_width="wrap_content"
                 android:layout_height="wrap_content"
                 android:layout_gravity="center"
                 android:layout_marginVertical="15dp"
                 android:fontFamily="@font/nexa_bold"
-                android:text="Logout"
+                android:text="@string/logout"
                 android:textColor="@color/blue"
                 android:textSize="20sp" />
         </androidx.cardview.widget.CardView>
