+++ app/src/main/res/layout/activity_search.xml
@@ -15,6 +15,7 @@
         android:layout_marginStart="@dimen/_8sdp"
         android:background="?android:attr/selectableItemBackground"
         android:clickable="true"
+        android:contentDescription="@string/picture"
         android:focusable="true"
         android:scaleType="fitCenter"
         android:src="@drawable/ic_baseline_arrow_back_24"
@@ -67,6 +68,7 @@
                 android:layout_centerVertical="true"
                 android:background="?android:attr/selectableItemBackground"
                 android:clickable="true"
+                android:contentDescription="@string/picture"
                 android:focusable="true"
                 android:scaleType="fitCenter"
                 android:src="@drawable/ic_baseline_search_24"
@@ -143,16 +145,12 @@
         android:id="@+id/searchPageNoItemsSearchedText"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:fontFamily="@font/lato"
+        android:layout_centerInParent="true"
+        android:fontFamily="@font/montserrat"
         android:text="@string/no_items_searched"
         android:textColor="@color/lightGrey"
         android:textSize="@dimen/_14sdp"
-        android:visibility="gone"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintHorizontal_bias="0.5"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent" />
+        android:visibility="visible" />
 
     <androidx.recyclerview.widget.RecyclerView
         android:id="@+id/searchPageRecyclerView"
@@ -162,11 +160,10 @@
         android:layout_marginTop="16dp"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintHorizontal_bias="0.0"
+        android:scrollbars="vertical"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toBottomOf="@+id/searchPageSection2"
-        tools:listitem="@layout/search_page_recycler_item">
-
-    </androidx.recyclerview.widget.RecyclerView>
+        tools:listitem="@layout/search_page_recycler_item" />
 
     <ProgressBar
         android:id="@+id/searchPageProgressBar"
@@ -174,9 +171,9 @@
         android:layout_height="wrap_content"
         android:layout_centerInParent="true"
         android:indeterminate="true"
-        android:visibility="gone"
         android:indeterminateTint="@color/white"
         android:minWidth="@dimen/_50sdp"
         android:minHeight="@dimen/_50sdp"
+        android:visibility="gone"
         tools:targetApi="lollipop" />
 </RelativeLayout>
