+++ app/src/main/res/layout/appwidget_recycler_item.xml
@@ -1,23 +1,23 @@
 <?xml version="1.0" encoding="utf-8"?>
 
 <RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    android:id="@+id/appWidgetRecyclerItem"
     android:layout_width="match_parent"
     android:layout_height="wrap_content"
     android:layout_marginVertical="@dimen/_12sdp"
-    android:background="@drawable/appwidget_outer_bg">
+    android:layout_marginStart="@dimen/_1sdp">
 
     <RelativeLayout
         android:id="@+id/appwidgetRecyclerViewItemSection1"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
         android:layout_marginTop="@dimen/_5sdp"
-        android:layout_marginStart="@dimen/_10sdp"
         android:layout_marginBottom="@dimen/_10sdp">
 
         <ImageView
             android:id="@+id/appwidgetRecyclerViewItemProfilePhoto"
-            android:layout_width="@dimen/_30sdp"
-            android:layout_height="@dimen/_30sdp"
+            android:layout_width="@dimen/_24sdp"
+            android:layout_height="@dimen/_24sdp"
             android:layout_centerVertical="true"
             android:layout_marginEnd="10dp"
             android:src="@drawable/ic_baseline_person_24" />
@@ -40,8 +40,7 @@
     <RelativeLayout
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
-        android:layout_below="@+id/appwidgetRecyclerViewItemSection1"
-        android:layout_marginStart="@dimen/_10sdp">
+        android:layout_below="@+id/appwidgetRecyclerViewItemSection1">
 
         <RelativeLayout
             android:id="@+id/appwidgetEventTypeIconRL"
@@ -56,7 +55,8 @@
                 android:layout_width="@dimen/_12sdp"
                 android:layout_height="@dimen/_12sdp"
                 android:layout_centerInParent="true"
-                android:layout_margin="@dimen/_10sdp"
+                android:layout_margin="@dimen/_6sdp"
+                android:contentDescription="@string/picture"
                 android:scaleType="fitXY"
                 android:src="@drawable/ic_github_fork" />
         </RelativeLayout>
@@ -66,7 +66,7 @@
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
             android:layout_marginEnd="@dimen/_10sdp"
-            android:layout_marginBottom="@dimen/_10sdp"
+            android:layout_marginBottom="0dp"
             android:layout_toEndOf="@+id/appwidgetEventTypeIconRL"
             android:background="@drawable/appwidget_inner_bg">
 
@@ -74,37 +74,45 @@
                 android:id="@+id/appwidgetRecyclerViewItemRelativeLayout"
                 android:layout_width="match_parent"
                 android:layout_height="wrap_content"
-                android:layout_margin="@dimen/_8sdp"
-                android:padding="@dimen/_6sdp">
+                android:layout_margin="@dimen/_4sdp"
+                android:padding="@dimen/_4sdp">
 
-                <TextView
-                    android:id="@+id/appwidgetRecyclerViewItemRepoName"
+                <RelativeLayout
+                    android:id="@+id/appwidgetRecyclerViewItemRepoNameRL"
                     android:layout_width="wrap_content"
-                    android:layout_height="wrap_content"
-                    android:fontFamily="@font/nexa_bold"
-                    android:text="@string/repo_name"
-                    android:textColor="@color/white" />
+                    android:layout_height="wrap_content">
+
+                    <TextView
+                        android:id="@+id/appwidgetRecyclerViewItemRepoName"
+                        android:layout_width="wrap_content"
+                        android:layout_height="wrap_content"
+                        android:fontFamily="@font/nexa_bold"
+                        android:text="@string/repo_name"
+                        android:textColor="@color/white"
+                        android:textSize="@dimen/_11sdp" />
+
+                    <TextView
+                        android:id="@+id/appwidgetRecyclerViewItemDate"
+                        android:layout_width="wrap_content"
+                        android:layout_height="wrap_content"
+                        android:layout_alignParentEnd="true"
+                        android:layout_marginStart="@dimen/_4sdp"
+                        android:text="@string/date"
+                        android:textColor="@color/lightGrey"
+                        android:textSize="@dimen/_8sdp" />
+                </RelativeLayout>
 
                 <TextView
                     android:id="@+id/appwidgetRecyclerViewItemMessage"
                     android:layout_width="wrap_content"
                     android:layout_height="wrap_content"
-                    android:layout_below="@+id/appwidgetRecyclerViewItemRepoName"
-                    android:layout_marginTop="@dimen/_10sdp"
+                    android:layout_below="@+id/appwidgetRecyclerViewItemRepoNameRL"
+                    android:layout_marginTop="@dimen/_5sdp"
                     android:fontFamily="@font/montserrat"
                     android:maxLines="3"
                     android:text="@string/appwidget_commit_message"
-                    android:textColor="@color/lightGrey" />
-
-                <TextView
-                    android:id="@+id/appwidgetRecyclerViewItemDate"
-                    android:layout_width="wrap_content"
-                    android:layout_height="wrap_content"
-                    android:layout_below="@id/appwidgetRecyclerViewItemMessage"
-                    android:layout_marginTop="@dimen/_5sdp"
-                    android:text="@string/date"
                     android:textColor="@color/lightGrey"
-                    android:textSize="@dimen/_8sdp" />
+                    android:textSize="@dimen/_10sdp" />
             </RelativeLayout>
         </RelativeLayout>
     </RelativeLayout>
