+++ app/src/main/res/layout/feed_page_recycler_item.xml
@@ -1,17 +1,17 @@
 <?xml version="1.0" encoding="utf-8"?>
 <androidx.cardview.widget.CardView xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
+    android:id="@+id/feedPageRecyclerViewItem"
     android:layout_width="match_parent"
     android:layout_height="wrap_content"
-    android:id="@+id/feedPageRecyclerViewItem"
     android:layout_marginHorizontal="@dimen/_10sdp"
     android:layout_marginVertical="@dimen/_4sdp"
-    app:cardBackgroundColor="@android:color/transparent"
-    android:foreground="?android:attr/selectableItemBackground"
     android:clickable="true"
+    android:focusable="true"
+    android:foreground="?android:attr/selectableItemBackground"
+    app:cardBackgroundColor="@android:color/transparent"
     app:cardCornerRadius="20dp"
-    app:cardElevation="0dp"
-    android:focusable="true">
+    app:cardElevation="0dp">
 
     <RelativeLayout
         android:layout_width="match_parent"
@@ -32,16 +32,18 @@
                 android:layout_width="wrap_content"
                 android:layout_height="wrap_content"
                 android:layout_marginEnd="@dimen/_8sdp"
-                app:cardBackgroundColor="@android:color/transparent"
                 android:clickable="true"
-                app:cardCornerRadius="30dp"
+                android:focusable="true"
                 android:foreground="?selectableItemBackground"
-                android:focusable="true">
+                android:shape="ring"
+                app:cardBackgroundColor="@android:color/transparent"
+                app:cardCornerRadius="@dimen/_15sdp">
 
                 <ImageView
                     android:id="@+id/feedPageRecyclerViewItemProfilePhoto"
                     android:layout_width="@dimen/_30sdp"
                     android:layout_height="@dimen/_30sdp"
+                    android:contentDescription="@string/picture"
                     android:src="@drawable/ic_baseline_person_24" />
             </androidx.cardview.widget.CardView>
 
@@ -52,6 +54,9 @@
                 android:layout_centerVertical="true"
                 android:layout_marginStart="@dimen/_10sdp"
                 android:layout_toEndOf="@+id/feedPageRecyclerViewItemProfilePhotoCard"
+                android:background="?selectableItemBackground"
+                android:clickable="true"
+                android:focusable="true"
                 android:fontFamily="@font/montserrat"
                 android:maxLines="1"
                 android:text="john doe"
@@ -72,15 +77,18 @@
                 android:layout_height="wrap_content"
                 android:layout_centerVertical="true"
                 android:layout_marginEnd="@dimen/_20sdp"
+                android:shape="ring"
                 app:cardBackgroundColor="@color/feedPageRecyclerItemColor"
-                app:cardCornerRadius="30dp">
+                app:cardCornerRadius="@dimen/_16sdp">
 
                 <ImageView
                     android:id="@+id/feedPageEventTypeIcon"
                     android:layout_width="@dimen/_16sdp"
                     android:layout_height="@dimen/_16sdp"
                     android:layout_margin="@dimen/_8sdp"
-                    android:src="@drawable/ic_github_fork" />
+                    android:contentDescription="@string/picture"
+                    android:src="@drawable/ic_github_fork"
+                    app:tint="@color/white" />
             </androidx.cardview.widget.CardView>
 
             <androidx.cardview.widget.CardView
